requirements.txt
Flask==0.10
python-instagram==0.8